+++ src/main/java/sparta/streaming/user/OAuth2UserService.java
@@ -1,6 +1,5 @@
 package sparta.streaming.user;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -9,8 +8,8 @@ import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
 import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
-import sparta.streaming.domain.CustomOAuth2User;
-import sparta.streaming.domain.User;
+import sparta.streaming.domain.user.CustomOAuth2User;
+import sparta.streaming.domain.user.User;
 
 import java.util.Map;
