+++ src/main/java/sparta/streaming/user/UserCertificationRepository.java
@@ -2,7 +2,7 @@ package sparta.streaming.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import sparta.streaming.domain.UserCertification;
+import sparta.streaming.domain.user.UserCertification;
 
 @Repository
 public interface UserCertificationRepository extends JpaRepository<UserCertification, Long> {
