+++ src/main/java/sparta/streaming/user/UserRepository.java
@@ -1,8 +1,6 @@
 package sparta.streaming.user;
 
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-import sparta.streaming.domain.User;
+import sparta.streaming.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
@@ -12,12 +10,4 @@ public interface UserRepository extends JpaRepository<User, Long> { // 엔티티
     Optional<User> findByEmail(String email);
 
     User findByUserId(Long userId);
-
-//    @Query("SELECT CASE WHEN COUNT(u) > 0 THEN TRUE ELSE FALSE END FROM User u WHERE u.email = :email")
-//    boolean existsByEmail(@Param("email") String email);
-//
-//    @Query("SELECT CASE WHEN COUNT(u) > 0 THEN TRUE ELSE FALSE END FROM User u WHERE u.userName = :userName")
-//    boolean existsByUserName(@Param("userName") String userName);
-
-
 }
