+++ src/main/java/sparta/streaming/user/UserService.java
@@ -1,25 +1,15 @@
 package sparta.streaming.user;
 
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.apache.coyote.BadRequestException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import sparta.streaming.domain.User;
-import sparta.streaming.dto.ResponseMessage;
+import sparta.streaming.domain.user.User;
 import sparta.streaming.dto.user.CreateUserRequestDto;
 import sparta.streaming.dto.user.PutUserRequestDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import sparta.streaming.dto.user.UserCommonDto;
-import sparta.streaming.handler.OAuth2SuccessHandler;
 import sparta.streaming.user.provider.JwtProvider;
 
 import java.util.List;
